requirements.txt
bibtexparser==1.4.3
pyparsing==3.2.5
pyperclip==1.11.0
PyYAML==6.0.3
